RF: Moved code around and cleaned up
